requirements.txt
wheel
jinja2
wget
python-generate-mac
natsort
typer[all]
pyyaml
rich
invoke
